refactor: Modify repo structure.
