+++ src/ansys/fluent/core/__init__.py
@@ -33,12 +33,12 @@ from ansys.fluent.core.logger import *
 
 # isort: on
 
+from ansys.fluent.core.exceptions import *
 from ansys.fluent.core.fields.field_data_interfaces import *
 from ansys.fluent.core.get_build_details import *
 from ansys.fluent.core.launcher.launch_options import *
 from ansys.fluent.core.launcher.launcher import *
-from ansys.fluent.core.parametric import *
-from ansys.fluent.core.pyfluent_warnings import *
+from ansys.fluent.core.legacy.local_parametric_study import *
 from ansys.fluent.core.search import *
 from ansys.fluent.core.services.batch_ops import *
 from ansys.fluent.core.session import *
@@ -69,9 +69,15 @@ if _os.path.exists(_README_FILE):
     with open(_README_FILE, encoding="utf8") as f:
         __doc__ = f.read()
 
+from ansys.fluent.core.legacy import (  # noqa: E402
+    local_parametric_study as _local_parametric_study,
+)
+from ansys.fluent.core.legacy import rpvars as _rpvars  # noqa: E402
 from ansys.fluent.core.session import file as _session_file  # noqa: E402
 
 _sys.modules["ansys.fluent.core.file_session"] = _session_file
+_sys.modules["ansys.fluent.core.rpvars"] = _rpvars
+_sys.modules["ansys.fluent.core.parametric"] = _local_parametric_study
 
 
 class Fluent(BaseSession):
