+++ src/ansys/fluent/core/fields/field_data_interfaces.py
@@ -833,7 +833,7 @@ class _ReturnFieldData:
                             surface_ids[count]
                         ][SurfaceDataType.FacesConnectivity.value]
                     else:
-                        from ansys.fluent.core.pyfluent_warnings import (
+                        from ansys.fluent.core.exceptions import (
                             PyFluentDeprecationWarning,
                         )
 
@@ -886,7 +886,7 @@ class _ReturnFieldData:
             if flatten_connectivity:
                 lines_data = pathlines_data[surface_ids[count]]["lines"]
             else:
-                from ansys.fluent.core.pyfluent_warnings import (
+                from ansys.fluent.core.exceptions import (
                     PyFluentDeprecationWarning,
                 )
