+++ src/ansys/fluent/core/fields/live_field_data.py
@@ -48,7 +48,10 @@ import weakref
 
 import numpy as np
 
-from ansys.fluent.core.exceptions import DisallowedValuesError
+from ansys.fluent.core.exceptions import (
+    DisallowedValuesError,
+    PyFluentDeprecationWarning,
+)
 from ansys.fluent.core.fields.field_data_interfaces import (
     BaseFieldDataSource,
     BaseFieldInfo,
@@ -70,7 +73,6 @@ from ansys.fluent.core.fields.field_data_interfaces import (
     _VectorFields,
     get_surfaces_from_objects,
 )
-from ansys.fluent.core.pyfluent_warnings import PyFluentDeprecationWarning
 from ansys.fluent.core.variable_strategies import (
     FluentFieldDataNamingStrategy as naming_strategy,
 )
