+++ src/ansys/fluent/core/fluent_connection.py
@@ -44,6 +44,7 @@ import weakref
 from deprecated.sphinx import deprecated
 import grpc
 
+from ansys.fluent.core.exceptions import InsecureGrpcWarning
 from ansys.fluent.core.launcher.error_warning_messages import (
     ALLOW_REMOTE_HOST_NOT_PROVIDED_IN_REMOTE,
     CERTIFICATES_FOLDER_NOT_PROVIDED_AT_CONNECT,
@@ -52,7 +53,6 @@ from ansys.fluent.core.launcher.error_warning_messages import (
 )
 from ansys.fluent.core.launcher.launcher_utils import ComposeConfig
 from ansys.fluent.core.module_config import config
-from ansys.fluent.core.pyfluent_warnings import InsecureGrpcWarning
 from ansys.fluent.core.services._protocols import ServiceProtocol
 from ansys.fluent.core.utils.execution import timeout_exec, timeout_loop
 from ansys.fluent.core.utils.file_transfer_service import ContainerFileTransferStrategy
