+++ src/ansys/fluent/core/session/base_meshing.py
@@ -41,10 +41,9 @@ import warnings
 
 from ansys.fluent.core._types import PathType
 from ansys.fluent.core.data_model_cache import DataModelCache, NameKey
-from ansys.fluent.core.exceptions import BetaFeaturesNotEnabled
+from ansys.fluent.core.exceptions import BetaFeaturesNotEnabled, PyFluentUserWarning
 from ansys.fluent.core.fluent_connection import FluentConnection
 from ansys.fluent.core.module_config import config
-from ansys.fluent.core.pyfluent_warnings import PyFluentUserWarning
 from ansys.fluent.core.services.scheme_interpreter import SchemeInterpreter
 from ansys.fluent.core.session._shared import (
     _make_datamodel_module,
