+++ src/ansys/fluent/core/session/session.py
@@ -53,13 +53,13 @@ if TYPE_CHECKING:
     )
     from ansys.fluent.core.launcher.pim_launcher import PIMArgsWithoutMode
 
-from ansys.fluent.core.fluent_connection import FluentConnection
-from ansys.fluent.core.journaling import Journal
-from ansys.fluent.core.pyfluent_warnings import (
+from ansys.fluent.core.exceptions import (
     PyFluentDeprecationWarning,
     PyFluentUserWarning,
 )
-from ansys.fluent.core.rpvars import RPVars
+from ansys.fluent.core.fluent_connection import FluentConnection
+from ansys.fluent.core.journaling import Journal
+from ansys.fluent.core.legacy.rpvars import RPVars
 from ansys.fluent.core.services.scheme_interpreter import SchemeInterpreter
 from ansys.fluent.core.utils.deprecate import deprecate_function
 from ansys.fluent.core.utils.fluent_version import FluentVersion
