+++ src/ansys/fluent/core/streaming_services/events_streaming.py
@@ -30,8 +30,7 @@ import logging
 from typing import Generic, TypeVar
 import warnings
 
-from ansys.fluent.core.exceptions import InvalidArgument
-from ansys.fluent.core.pyfluent_warnings import PyFluentDeprecationWarning
+from ansys.fluent.core.exceptions import InvalidArgument, PyFluentDeprecationWarning
 from ansys.fluent.core.streaming_services._events_info_store import (
     AboutToInitializeSolutionEventInfo,
     AboutToLoadCaseEventInfo,
