+++ src/ansys/fluent/core/utils/file_transfer_service.py
@@ -31,7 +31,7 @@ import shutil
 from typing import Any, Protocol
 import warnings
 
-from ansys.fluent.core.pyfluent_warnings import PyFluentUserWarning
+from ansys.fluent.core.exceptions import PyFluentUserWarning
 from ansys.fluent.core.utils import get_user_data_dir
 from ansys.fluent.core.utils.deprecate import deprecate_arguments
 import ansys.platform.instancemanagement as pypim
