+++ src/ansys/fluent/core/workflow.py
@@ -33,7 +33,7 @@ import threading
 from typing import Any
 import warnings
 
-from ansys.fluent.core.pyfluent_warnings import (
+from ansys.fluent.core.exceptions import (
     PyFluentDeprecationWarning,
 )
 from ansys.fluent.core.services.object_model import (
