+++ tests/integration/test_optislang/test_optislang_integration.py
@@ -33,7 +33,7 @@ from test_utils import pytest_approx
 import ansys.fluent.core as pyfluent
 from ansys.fluent.core import examples
 from ansys.fluent.core.docker.utils import get_grpc_launcher_args_for_gh_runs
-from ansys.fluent.core.filereader.case_file import CaseFile
+from ansys.fluent.core.file_reader.case_file import CaseFile
 from ansys.fluent.core.utils.fluent_version import FluentVersion
