+++ tests/parametric/test_local_parametric_setup.py
@@ -22,7 +22,7 @@
 # SOFTWARE.
 
 from ansys.fluent.core import examples
-from ansys.fluent.core.parametric import LocalParametricStudy
+from ansys.fluent.core.legacy.local_parametric_study import LocalParametricStudy
 
 
 def test_local_parametric_setup():
