+++ tests/test_lispy.py
@@ -21,7 +21,7 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-from ansys.fluent.core.filereader import lispy
+from ansys.fluent.core.file_reader import lispy
 
 scm_pys = (
     ("()", []),
