+++ tests/test_settings_reader.py
@@ -24,7 +24,7 @@
 import pytest
 
 from ansys.fluent.core import examples
-from ansys.fluent.core.filereader.case_file import SettingsFile as SettingsReader
+from ansys.fluent.core.file_reader.case_file import SettingsFile as SettingsReader
 
 
 def call_settings_reader(
